+++ base/base/allocator_helpers.inl
@@ -11,18 +11,18 @@ inline Header *header(void *data)
     return (Header *)memory::pointer_sub(p, sizeof(Header));
 }
 
-inline void fill(Header *header, void *data, uint32 size)
+inline void fill(Header *header, void *data, uint64 size)
 {
     header->size = size;
     #if defined(TRACING)
         header->tracing_marker = TRACING_MARKER;
     #endif
 
-    auto p = (uint32 *)memory::pointer_add(header, sizeof(Header));
+    auto p = (uint64 *)memory::pointer_add(header, sizeof(Header));
     while (p < data)
         *p++ = HEADER_PAD_VALUE;
 }
 
-inline uint32 size_with_padding(uint32 size, uint32 align) {
+inline uint64 size_with_padding(uint64 size, uint32 align) {
     return size + align * 2 + sizeof(Header);
 }
